cycles now should not overflow the stack when printing, they still cause memory leaks though
