+++ cogs/collections/dll.scm
@@ -52,15 +52,12 @@
   (insert-head dlist 1)
   (displayln dlist)
   (insert-tail dlist 4)
-;   (displayln dlist)
+  (displayln dlist)
   (insert-after dlist (dllist-head dlist) 2)
-;   (displayln dlist)
+  (displayln dlist)
   (let* ((next-to-last (insert-before dlist (dllist-tail dlist) 3))
          (bad-link (insert-before dlist next-to-last 42)))
     (remove-link dlist bad-link))
-;   (displayln dlist)
+  (displayln dlist)
   (displayln (dllist-elements dlist))
-;   (displayln dlist)
-  
-  
-  )
+  (displayln dlist))
