+++ crates/steel/src/compiler/map.rs
@@ -24,12 +24,22 @@ impl SymbolMap {
         }
     }
 
+    pub fn len(&self) -> usize {
+        self.values.len()
+    }
+
     // pub fn add(&mut self, ident: &str) -> usize {
     //     let idx = self.values.len();
     //     self.values.push(ident.to_string());
     //     // println!("`add`: {} @ {}", ident, idx);
     //     idx
     // }
+
+    pub fn roll_back(&mut self, index: usize) {
+        for value in self.values.drain(index..) {
+            self.map.remove(&value);
+        }
+    }
 
     pub fn get_or_add(&mut self, ident: &str) -> usize {
         // let rev_iter = self.values.iter().enumerate().rev();
@@ -60,6 +70,9 @@ impl SymbolMap {
 
             self.map.insert(ident.to_string(), idx);
 
+            // Add the values so we can do a backwards resolution
+            self.values.push(ident.to_string());
+
             idx
         }
     }
@@ -83,54 +96,4 @@ impl SymbolMap {
             .copied()
             .ok_or_else(throw!(FreeIdentifier => ident.to_string()))
     }
-
-    // // TODO -> want to package the metadata up for declaring structs
-    // // into a program so that someone can take a binary and load it correctly
-    // // probably front load the program with a vector of strings declaring what struct functions will inevitably
-    // // be declared
-    // pub fn insert_struct_function_names_from_concrete<'a>(
-    //     &mut self,
-    //     struct_builder: &'a StructFuncBuilderConcrete,
-    // ) -> Vec<usize> {
-    //     let mut indices = Vec::new();
-
-    //     // Constructor
-    //     indices.push(self.get_or_add(&struct_builder.name));
-    //     // Predicate
-    //     indices.push(self.get_or_add(format!("{}?", &struct_builder.name).as_str()));
-    //     for field in &struct_builder.fields {
-    //         // Getter
-    //         indices.push(self.get_or_add(format!("{}-{}", &struct_builder.name, field).as_str()));
-    //         // Setter
-    //         indices
-    //             .push(self.get_or_add(format!("set-{}-{}!", &struct_builder.name, field).as_str()));
-    //     }
-
-    //     indices
-    // }
-
-    // // TODO -> want to package the metadata up for declaring structs
-    // // into a program so that someone can take a binary and load it correctly
-    // // probably front load the program with a vector of strings declaring what struct functions will inevitably
-    // // be declared
-    // pub fn insert_struct_function_names<'a>(
-    //     &mut self,
-    //     struct_builder: &'a StructFuncBuilder,
-    // ) -> Vec<usize> {
-    //     let mut indices = Vec::new();
-
-    //     // Constructor
-    //     indices.push(self.get_or_add(&struct_builder.name));
-    //     // Predicate
-    //     indices.push(self.get_or_add(format!("{}?", &struct_builder.name).as_str()));
-    //     for field in &struct_builder.fields {
-    //         // Getter
-    //         indices.push(self.get_or_add(format!("{}-{}", &struct_builder.name, field).as_str()));
-    //         // Setter
-    //         indices
-    //             .push(self.get_or_add(format!("set-{}-{}!", &struct_builder.name, field).as_str()));
-    //     }
-
-    //     indices
-    // }
 }
